vfs: Move security_inode_killpriv() after permission checks
euler inclusion
from: https://lists.debian.org/debian-kernel/2014/11/msg00320.html
target:    kernel 3.4
category:  CVE/CVE-2015-1350

This commit was not in mainline. It fixed a cve.

-------------------------------

chown() and write() should clear all privilege attributes on
a file - setuid, setgid, setcap and any other extended
privilege attributes.

However, any attributes beyond setuid and setgid are managed by the
LSM and not directly by the filesystem, so they cannot be set along
with the other attributes.

Currently we call security_inode_killpriv() in notify_change(),
but in case of a chown() this is too early - we have not called
inode_change_ok() or made any filesystem-specific permission/sanity
checks.

Add a new function setattr_killpriv() which calls
security_inode_killpriv() if necessary, and change the setattr()
implementation to call this in each filesystem that supports xattrs.
This assumes that extended privilege attributes are always stored in
xattrs.

Compile-tested only.

XXX This is a silent change to the VFS API, but we should probably
change something so OOT filesystems fail to compile if they aren't
updated to call setattr_killpriv().

Reported-by: Ben Harris <bjh21@cam.ac.uk>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
References: https://bugs.debian.org/770492
[xr: apply to 3.4:
 - adjust context
 - drop fix in fuse, because these code had been applied
 - fix some another caller in xfs]
Signed-off-by: Rui Xiang <rui.xiang@huawei.com>