KEYS: Don't permit request_key() to construct a new keyring
mainline inclusion
from: v4.15
target: kernel 3.4
category:  CVE-2017-17807

----------------------

commit 911b79cde95c7da0ec02f48105358a36636b7a71 upstream.

If request_key() is used to find a keyring, only do the search part - don't
do the construction part if the keyring was not found by the search.  We
don't really want keyrings in the negative instantiated state since the
rejected/negative instantiation error value in the payload is unioned with
keyring metadata.

Now the kernel gives an error:

	request_key("keyring", "#selinux,bdekeyring", "keyring", KEY_SPEC_USER_SESSION_KEYRING) = -1 EPERM (Operation not permitted)

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Jason Yan <yanaijie@huawei.com>
Reviewed-by: Li Kun <hw.likun@huawei.com>