md:raid1: fix a dead loop when read from a WriteMostly disk
euler inclusion
from:     hulk 4.1
target:    kernel 3.4
category:  bugfix

-------------------------------

If first_bad == this_sector when we get the WriteMostly disk
in read_balance(), valid disk will be returned with zero
max_sectors. It'll lead to a dead loop in make_request(), and
OOM will happen because of endless allocation of struct bio.

Since we can't get data from this disk in this case, so
continue for another disk.

Signed-off-by: Wei Fang <fangwei1@huawei.com>
Signed-off-by: Rui Xiang <rui.xiang@huawei.com>