ipv6: Introduce ip6_flow_hdr() to fill version, tclass and flowlabel.
mainline inclusion
from: linux tags/v3.9-rc1~139^2~398
target:    kernel 3.4
category:  feture/IPv6

-------------------------------

commit 3e4e4c1f2da66b29ee9379ca29f8dd620c2b5a1f upstream.

This is not only for readability but also for optimization.
What we do here is to build the 32bit word at the beginning of the ipv6
header (the "ip6_flow" virtual member of struct ip6_hdr in RFC3542) and
we do not need to read the tclass portion of the target buffer.

Signed-off-by: YOSHIFUJI Hideaki <yoshfuji@linux-ipv6.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Yufen Wang <wangyufen@huawei.com>