tcp: fix inet6_csk_route_req() for link-local addresses
mainline inclusion
from:      tags/v3.6-rc1~125^2~353
target:    kernel 3.4
category:  bugfix/bug 2209

-------------------------------

commit 9247869ee661b046510c19a36cf0d91d9c2639d3 upstream.

Fix inet6_csk_route_req() to use as the flowi6_oif the treq->iif,
which is correctly fixed up in tcp_v6_conn_request() to handle the
case of link-local addresses. This brings it in line with the
tcp_v6_send_synack() code, which is already correctly using the
treq->iif in this way.

Signed-off-by: Neal Cardwell <ncardwell@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Yufen Wang <wangyufen@huawei.com>