NVMe: check for integer overflow in nvme_map_user_pages()
Mainline inclusion
from: linux 3.12
target: kernel 3.4
category:  Bugfix
---------------------

upstream 5460fc03105fbed01fe27aa572d9f65bb410a61d commit

You need to have CAP_SYS_ADMIN to trigger this overflow but it makes the
static checkers complain so we should fix it.  The worry is that
"length" comes from copy_from_user() so we need to check that "length +
offset" can't overflow.

I also changed the min_t() cast to be unsigned instead of signed.  Now
that we cap "length" to INT_MAX it doesn't make a difference, but it's a
little easier for reviewers to know that large values aren't cast to
negative.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
Signed-off-by: Ding Xiang <dingxiang@huawei.com>