hugetlb: fix race between alloc and soft offline hugetlb
euler inclusion
category: bugfix
bugzilla: 200
DTS: DTS2016121900567

-----------------------------------------------

There is a race between alloc hugetlb and soft offline hugetlb.
The result is that we alloc a hugetlb which marked hwpoison flag,
also soft offline hugetlb failed but the the count increase. This
will trigger page fault handler to kill the process when the data
accessed.

__do_fault()
	PageHWPoison()
		goto uncharge_out;

Upstream v3.12(c8721bbb, mm: memory-hotplug: enable memory hotplug
to handle hugepage)has fix this bug, but it's a large patchset which
intorduces a new feature. So I just backport a part of that patch.

Signed-off-by: Xishi Qiu <qiuxishi@huawei.com>
Signed-off-by: Jason Yan <yanaijie@huawei.com>