scsi: avoid a permanent stop of the scsi device's request queue
mainline inclusion
from:      linux 3.10
target:    linux 3.4
category:  Bugfix/Bug 2629/DTS2016122000586

----------------------------------------

A race between scanning and fc_remote_port_delete() may result in a
permanent stop if the device gets blocked before scsi_sysfs_add_sdev()
and unblocked after.  The reason is that blocking a device sets both
the SDEV_BLOCKED state and the QUEUE_FLAG_STOPPED.  However,
scsi_sysfs_add_sdev() unconditionally sets SDEV_RUNNING which causes
the device to be ignored by scsi_target_unblock() and thus never have
its QUEUE_FLAG_STOPPED cleared leading to a device which is apparently
running but has a stopped queue.

We actually have two places where SDEV_RUNNING is set: once in
scsi_add_lun() which respects the blocked flag and once in
scsi_sysfs_add_sdev() which doesn't.  Since the second set is entirely
spurious, simply remove it to fix the problem.

This patch is a little bit different with the patch applied to the
upstream, it's because the patch applied to the upstream hasn't
thoroughly fixed this issue, see more about this:
http://hulk.huawei.com/bugzilla/show_bug.cgi?id=2629

Reported-by: Zengxi Chen <chenzengxi@huawei.com>
Signed-off-by: Wei Fang <fangwei1@huawei.com>
Signed-off-by: Jason Yan <yanaijie@huawei.com>