fs: ext3/ext4: increase the protection of nlink dec and inode destroy
mainline inclusion
from: NA
target: kernel 3.4
category:  DTS2017020903219

----------------------

Because of the disk and hardware issue, the ext3/4 filesystem have
many errors, the inode->i_nlink of ext3/4 becomes zero abnormally
but the dentry is still positive, it will cause memory corruption
after the following process:

 1) Due to the inode->i_nlink is 0, this inode will be added into
the orhpan list,
 2) ext4_rename() and ext3_rename() cover this inode, and drop_nlink()
will reverse the inode->i_nlink to 0xFFFFFFFF,
 3) iput() add this inode to LRU,
 4) evict() will call destroy_inode() to destroy this inode but
skip removing it from the orphan list,
 5) after this, the inode's memory address space will be used by
other module, when the ext3/4 filesystem change the orphan list, it will
trample other module's data and then may cause oops.

Although we cannot avoid hardware and disk errors, we can control the
softwore error in the ext3/4 module, do not affect other modules and
increase the difficulty of locating problems.

This patch avoid inode->i_nlink underflow and remove the inode from the
orphan list when destroy it if the list is not empty.

Signed-off-by: zhangyi <yi.zhang@hauwei.com>
Signed-off-by: Jason Yan <yanaijie@huawei.com>