md/bitmap: don't free pages and attrs when writing bitmap gets error
euler inclusion
category: bugfix
bugzilla:3923
DTS: DTS2017102308898
CVE: NA

-------------------------------------------------

When writing bitmap failed, filemap/filemap_attr/file_pages/sb_page
will be cleared and freed by bitmap_file_unmap(), and it will cause
oops like the following one:

BUG: unable to handle kernel NULL pointer dereference at 0000000000000010
IP: [<ffffffff81475b96>] bitmap_endwrite+0x116/0x250
[<ffffffff814609aa>] close_write+0x8a/0xa0
[<ffffffff814615d4>] r1_bio_write_done+0x24/0x50
[<ffffffff81464a37>] raid1_end_write_request+0xc7/0x250
[<ffffffff811d64b8>] bio_endio+0x18/0x30
[<ffffffff81315a6b>] req_bio_endio.isra.57+0xcb/0xe0
[<ffffffff81315d55>] blk_update_request+0x2d5/0x4a0
[<ffffffff81315f3b>] blk_update_bidi_request+0x1b/0x80
[<ffffffff81315fba>] blk_end_bidi_request+0x1a/0x60
[<ffffffff8131600b>] blk_end_request+0xb/0x10
[<ffffffff81429afb>] scsi_io_completion+0xab/0x770
[<ffffffff8141fd99>] scsi_finish_command+0xb9/0x120
[<ffffffff81429946>] scsi_softirq_done+0x126/0x150
[<ffffffff8131cf1b>] blk_done_softirq+0x7b/0x90
[<ffffffff81087220>] __do_softirq+0xd0/0x220
[<ffffffff81585d0c>] call_softirq+0x1c/0x30
[<ffffffff810430fd>] do_softirq+0x5d/0x90
[<ffffffff8108760d>] irq_exit+0xad/0xc0
[<ffffffff8106008f>] smp_call_function_single_interrupt+0x2f/0x40
[<ffffffff815857da>] call_function_single_interrupt+0x6a/0x70
<EOI>
[<ffffffff810496a0>] default_idle+0x50/0x1b0
[<ffffffff8104a2e6>] cpu_idle+0xa6/0xf0
[<ffffffff8156d494>] start_secondary+0x226/0x228

To fix it, we can add more checks for the freeing of the bitmap pages
and attrs, or we can remove the freeing of pages and attrs and check
the write failure accordingly when writing back the bitmap. The former
is tedious and is prone to a missed check, and the latter is simpler.

Based on 62f82fa ("md/bitmap: remove async freeing of bitmap file").

Signed-off-by: Hou Tao <houtao1@huawei.com>
Signed-off-by: Jason Yan <yanaijie@huawei.com>