PCI: fix the wrong class type for HiSilicon NP 5896 device
hulk inclusion
category: bugfix
bugzilla: 13666
CVE: NA

------------------------------

The following patch set the class type as 'PCI_BASE_CLASS_NETWORK'. But
'PCI_BASE_CLASS_NETWORK' is actually the higher 8 bits of the class type
of a network device. We should set it as 'PCI_CLASS_NETWORK_ETHERNET'.
This patch fixes it.

Fixes: ba8bc9c15d20 ("PCI: Add quirk for hisilicon NP devices 5896")
Signed-off-by: Xiongfeng Wang <wangxiongfeng2@huawei.com>
Reviewed-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>